pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ai_text_outline"
version = "0.1.0"
authors = [
  { name="OpenPecha", email="dev@openpecha.org" },
]
description = "Extract Table of Contents from Tibetan texts and return section start indices"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "rapidfuzz>=3.0.0",
]

[project.optional-dependencies]
gemini = ["google-generativeai>=0.5.0"]
openai = ["openai>=1.0.0"]
claude = ["anthropic>=0.20.0"]
all = [
    "google-generativeai>=0.5.0",
    "openai>=1.0.0",
    "anthropic>=0.20.0",
]
dev = [
    "pytest",
    "pytest-cov",
    "pre-commit",
]

[project.urls]
"Homepage" = "https://github.com/OpenPecha/ai-text-outline"
"Bug Tracker" = "https://github.com/OpenPecha/ai-text-outline/issues"
